restores login functionality after refactor
