+++ cosc360-rsvp/client/src/features/adminManagement/AdminUserManage.jsx
@@ -1,7 +1,7 @@
-import AdminSidebar from "../../client/components/AdminSidebar";
-import AdminManagement from "../../client/components/AdminManagement"
-import topNav from "../../client/components/topNav";
-import '../css/adminManagement.css';
+import AdminSidebar from "../../components/AdminSidebar";
+import AdminManagement from "./AdminManagement"
+import topNav from "../../components/topNav";
+import './adminManagement.css';
 
 function AdminUserManage(){
     return(
