+++ cosc360-rsvp/client/src/features/event/EventContainer.jsx
@@ -1,4 +1,4 @@
-import '../../css/Event.css'
+import './Event.css'
 import EventDetails from './EventDetails';
