+++ cosc360-rsvp/client/src/features/event/event.jsx
@@ -1,8 +1,8 @@
-import ReviewContainer from "../../client/components/event/review_bar/ReviewContainer";
-import EventContainer from "../../client/components/event/EventContainer";
-import TopNav from "../../client/components/topNav";
-import Sidebar from "../../client/components/sidebar";
-import '../css/Event.css';
+import ReviewContainer from "./review_bar/ReviewContainer";
+import EventContainer from "./EventContainer";
+import TopNav from "../../components/topNav";
+import Sidebar from "../../components/sidebar";
+import './Event.css';
 
 const event = {
     title: "Graduation Party",
