+++ cosc360-rsvp/client/src/features/event/review_bar/ReviewContainer.jsx
@@ -1,4 +1,4 @@
-import '../../../css/Event.css';
+import '../Event.css';
 import ReviewItem from './ReviewItem.jsx';
 
 function getAverageRating(reviews) {
