+++ cosc360-rsvp/client/src/features/login/LoginCard.css
@@ -70,7 +70,7 @@
 
 body, html {
     margin: 0;
-    background-image: url("/src/assets/login-background.jpg");
+    background-image: url("/src/components/assets/login-background.jpg");
     background-repeat: repeat;
     background-size: 300px 300px;
     justify-content:center;
