+++ cosc360-rsvp/client/src/features/login/api/Login.js
@@ -1,4 +1,4 @@
-import { apiClient } from "../../../lib/api-clients.js";
+import { apiClient } from "../../../lib/api-client.js";
 
 export async function loginApi(username, password){
     const data = await apiClient("/login", {
