+++ cosc360-rsvp/client/src/features/register/RegisterForm.jsx
@@ -1,5 +1,5 @@
 import {useState} from "react";
-import "../css/App.css";
+import "../../app/App.css";
 
 
 function RegisterForm(){
