+++ cosc360-rsvp/client/src/main.jsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter } from "react-router-dom";
 import './css/index.css'
-import App from './App.jsx'
+import App from './app/App.jsx'
 
 /* Components don't go here! Put them in App.jsx */
