+++ cosc360-rsvp/client/src/pages/myEvents.jsx
@@ -1,7 +1,7 @@
-import EventContainer from "../../client/components/event_cards/EventContainer";
-import Sidebar from "../../client/components/sidebar";
-import AdminSidebar from "../../client/components/AdminSidebar";
-import TopNav from "../../client/components/topNav";
+import EventContainer from "../features/event/event_cards/EventContainer";
+import Sidebar from "../components/sidebar";
+import AdminSidebar from "../components/AdminSidebar";
+import TopNav from "../components/topNav";
 
 
 const username = "Lexi Loudiadis"
