+++ cosc360-rsvp/server/src/modules/services/processLogin.js
@@ -1,7 +1,7 @@
 import express from "express";
 const router = express.Router();
 
-router.post("api/login", (req, res) => {
+router.post("/login", (req, res) => {
   const { username, password } = req.body;
   if (username === "admin" && password === "1234") {
     res.json({ message: "Login successful" });
