+++ cosc360-rsvp/vite.config.js
@@ -3,6 +3,7 @@ import react from '@vitejs/plugin-react'
 
 // https://vite.dev/config/
 export default defineConfig({
+  root: "client",
   plugins: [react()],
   
   server: {
